# Refactor memap for speed