Improves performance by 10-15x on my machine. See below for individual compiler comparisons.

Resolves #4976
Resolves #5124
Resolves #5127 

Todo:
 - [x] Parsing tests (let's never let this non-performance bug in again)
 - [x] Resolve the `Malformed input...` output from GCC