@theotherjimmy 2017-09-15
# GCC Performance
## Before
```
# git checkout master 
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
# time python2 tools/memap.py -t GCC_ARM BUILD/tests/nrf51_dk/gcc_arm/TESTS/events/queue/queue.map -d 0
+------------------+-------+-------+------+
| Module           | .text | .data | .bss |
+------------------+-------+-------+------+
| [fill]           |    90 |     3 |   41 |
| [lib]/libc.a     | 24469 |  2472 |   56 |
| [lib]/libgcc.a   |  7134 |     0 |    0 |
| [lib]/libnosys.a |    32 |     0 |    0 |
| [lib]/misc       |   208 |     8 |   25 |
| blinky/BUILD     | 20425 |   261 | 5324 |
| fat-fs/BUILD     |  1389 |     0 | 1644 |
| main.o           |  9751 |     4 |  413 |
| mbed-os/BUILD    |  9215 |    12 |  769 |
| Subtotals        | 72713 |  2760 | 8272 |
+------------------+-------+-------+------+
Total Static RAM memory (data + bss): 11032 bytes
Total Flash memory (text + data): 75473 bytes

real    1.012522411s
```
## After
```
# git checkout improve-memap-performance
Switched to branch 'improve-memap-performance'
Your branch is up-to-date with 'theotherjimmy/improve-memap-performance'.
# time python2 tools/memap.py -t GCC_ARM BUILD/tests/nrf51_dk/gcc_arm/TESTS/events/queue/queue.map -d 0
Malformed input found when parsing GCC map: linker stubs
Malformed input found when parsing GCC map: linker stubs
Malformed input found when parsing GCC map: linker stubs
Malformed input found when parsing GCC map: linker stubs
+--------------------------------+-------+-------+------+
| Module                         | .text | .data | .bss |
+--------------------------------+-------+-------+------+
| TESTS/events                   |  9751 |     4 |  413 |
| [fill]                         |    90 |     3 |   41 |
| [lib]/libc.a                   | 24469 |  2472 |   56 |
| [lib]/libgcc.a                 |  7134 |     0 |    0 |
| [lib]/libnosys.a               |    32 |     0 |    0 |
| [lib]/misc                     |   208 |     8 |   25 |
| drivers/I2C.o                  |    18 |     0 |    0 |
| drivers/InterruptIn.o          |    90 |     0 |    0 |
| drivers/RawSerial.o            |   392 |     0 |    0 |
| drivers/SerialBase.o           |   398 |     0 |    0 |
| drivers/Ticker.o               |   238 |     0 |    0 |
| drivers/Timeout.o              |   167 |     0 |    0 |
| drivers/Timer.o                |   202 |     0 |    0 |
| drivers/TimerEvent.o           |   180 |     0 |    0 |
| events/EventQueue.o            |    62 |     0 |    0 |
| events/equeue                  |  1514 |     0 |   93 |
| features/frameworks            |  7286 |    69 |  513 |
| hal/mbed_gpio.o                |   220 |     0 |    0 |
| hal/mbed_lp_ticker_api.o       |    56 |     0 |   24 |
| hal/mbed_sleep_manager.o       |   219 |     0 |    2 |
| hal/mbed_ticker_api.o          |   436 |     0 |    0 |
| hal/mbed_us_ticker_api.o       |    56 |     0 |   24 |
| platform/mbed_alloc_wrappers.o |    44 |     0 |    0 |
| platform/mbed_assert.o         |    85 |     0 |    0 |
| platform/mbed_board.o          |   366 |     0 |    0 |
| platform/mbed_critical.o       |    62 |     0 |    0 |
| platform/mbed_error.o          |    44 |     0 |    1 |
| platform/mbed_retarget.o       |   647 |     4 |  264 |
| platform/mbed_wait_api_rtos.o  |    64 |     0 |    0 |
| rtos/TARGET_CORTEX             | 10647 |   180 | 6036 |
| rtos/Thread.o                  |     8 |     0 |    0 |
| targets/TARGET_NORDIC          |  7528 |    20 |  780 |
| Subtotals                      | 72713 |  2760 | 8272 |
+--------------------------------+-------+-------+------+
Total Static RAM memory (data + bss): 11032 bytes
Total Flash memory (text + data): 75473 bytes

real    0.093222451s
```