@theotherjimmy 2017-09-15
# Arm Compiler 5/6 performance

## Before
```
# git checkout master 
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
# time python2 tools/memap.py -t ARM BUILD/tests/nrf51_dk/arm/TESTS/events/queue/queue.map -e table
+-----------------+-------+-------+------+
| Module          | .text | .data | .bss |
+-----------------+-------+-------+------+
| [lib]/c_p.l     | 11603 |    16 |  348 |
| [lib]/cpprt_p.l |    44 |     0 |    0 |
| [lib]/fz_ps.l   |    32 |     0 |    0 |
| [lib]/m_ps.l    |    44 |     0 |    0 |
| anon$$obj.o     |    32 |     0 |    0 |
| blinky/BUILD    | 22684 |   395 | 5115 |
| fat-fs/BUILD    |  1680 |  2369 |   76 |
| main.o          | 10902 |     8 |  408 |
| mbed-os/BUILD   | 10677 |    74 | 1104 |
| Subtotals       | 57698 |  2862 | 7051 |
+-----------------+-------+-------+------+
Total Static RAM memory (data + bss): 9913 bytes
Total Flash memory (text + data): 60560 bytes

real    0.884907055s
```

## After
```
# git checkout improve-memap-performance
Switched to branch 'improve-memap-performance'
Your branch is up-to-date with 'theotherjimmy/improve-memap-performance'.
# time python2 tools/memap.py -t ARM BUILD/tests/nrf51_dk/arm/TESTS/events/queue/queue.map -e table
+--------------------------------+-------+-------+------+
| Module                         | .text | .data | .bss |
+--------------------------------+-------+-------+------+
| ./main.o                       | 10902 |     8 |  408 |
| [lib]/c_p.l                    | 11603 |    16 |  348 |
| [lib]/cpprt_p.l                |    44 |     0 |    0 |
| [lib]/fz_ps.l                  |    32 |     0 |    0 |
| [lib]/m_ps.l                   |    44 |     0 |    0 |
| anon$$obj.o                    |    32 |     0 |    0 |
| drivers/I2C.o                  |    76 |     0 |    0 |
| drivers/InterruptIn.o          |    85 |     0 |    0 |
| drivers/RawSerial.o            |   324 |     0 |    0 |
| drivers/SerialBase.o           |   460 |     0 |    0 |
| drivers/Ticker.o               |   198 |     0 |    0 |
| drivers/Timeout.o              |    64 |     0 |    0 |
| drivers/Timer.o                |   188 |     0 |    0 |
| drivers/TimerEvent.o           |   144 |     0 |    0 |
| events/EventQueue.o            |    76 |     0 |    0 |
| events/equeue                  |  1838 |     8 |   88 |
| features/frameworks            |  8049 |   140 |  512 |
| hal/mbed_gpio.o                |   256 |     0 |    0 |
| hal/mbed_lp_ticker_api.o       |    56 |     0 |   24 |
| hal/mbed_sleep_manager.o       |   216 |     2 |    0 |
| hal/mbed_ticker_api.o          |   464 |     0 |    0 |
| hal/mbed_us_ticker_api.o       |    56 |     0 |   24 |
| platform/FileBase.o            |   244 |     4 |   52 |
| platform/FilePath.o            |   170 |     0 |    0 |
| platform/mbed_alloc_wrappers.o |    16 |     0 |    0 |
| platform/mbed_assert.o         |    84 |     0 |    0 |
| platform/mbed_board.o          |   380 |     0 |    0 |
| platform/mbed_critical.o       |    64 |     0 |    0 |
| platform/mbed_error.o          |    44 |     1 |    0 |
| platform/mbed_retarget.o       |  1191 |     8 |  376 |
| platform/mbed_wait_api_rtos.o  |    56 |     0 |    0 |
| rtos/Mutex.o                   |   252 |     0 |    0 |
| rtos/TARGET_CORTEX             | 11987 |  2569 | 4300 |
| rtos/Thread.o                  |     8 |     0 |    0 |
| targets/TARGET_NORDIC          |  7995 |   106 |  919 |
| Subtotals                      | 57698 |  2862 | 7051 |
+--------------------------------+-------+-------+------+
Total Static RAM memory (data + bss): 9913 bytes
Total Flash memory (text + data): 60560 bytes

real    0.060380164s
```