@theotherjimmy 2017-09-15
# IAR Performance

## Before 
```
# git checkout master 
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
# time python2 tools/memap.py -t IAR BUILD/tests/nrf51_dk/iar/TESTS/events/queue/queue.map
+----------------------+-------+-------+------+
| Module               | .text | .data | .bss |
+----------------------+-------+-------+------+
| [lib]/dl6M_tlf.a     | 11882 |   480 |  716 |
| [lib]/dlpp6M_tl_fc.a |    64 |     0 |    0 |
| [lib]/m6M_tl.a       |  1368 |     0 |    0 |
| [lib]/rt6M_tl.a      |   974 |     0 |    0 |
| [misc]               |   239 |     0 |    0 |
| blinky/BUILD         | 20712 |   488 | 5514 |
| fat-fs/BUILD         |  1520 |     0 | 1700 |
| main.o               | 13206 |     0 |  413 |
| mbed-os/BUILD        |  8870 |     8 |  818 |
| Subtotals            | 58835 |   976 | 9161 |
+----------------------+-------+-------+------+
Total Static RAM memory (data + bss): 10137 bytes
Total Flash memory (text + data): 59811 bytes

real    0.757136647s
```

## After
```
git checkout improve-memap-performance
Switched to branch 'improve-memap-performance'
Your branch is up-to-date with 'theotherjimmy/improve-memap-performance'.
# time python2 tools/memap.py -t IAR BUILD/tests/nrf51_dk/iar/TESTS/events/queue/queue.map
+--------------------------------+-------+-------+------+
| Module                         | .text | .data | .bss |
+--------------------------------+-------+-------+------+
| ./main.o                       | 13206 |     0 |  413 |
| [lib]/dl6M_tlf.a               | 11882 |   480 |  716 |
| [lib]/dlpp6M_tl_fc.a           |    64 |     0 |    0 |
| [lib]/m6M_tl.a                 |  1368 |     0 |    0 |
| [lib]/rt6M_tl.a                |   974 |     0 |    0 |
| [misc]                         |   239 |     0 |    0 |
| drivers/RawSerial.o            |   276 |     0 |    0 |
| drivers/SerialBase.o           |   420 |     0 |    0 |
| drivers/Ticker.o               |   206 |     0 |    0 |
| drivers/Timeout.o              |    64 |     0 |    0 |
| drivers/Timer.o                |   210 |     0 |    0 |
| drivers/TimerEvent.o           |   146 |     0 |    0 |
| events/EventQueue.o            |    72 |     0 |    0 |
| events/equeue                  |  1556 |     0 |   96 |
| features/frameworks            |  7076 |   196 |  356 |
| hal/mbed_gpio.o                |   220 |     0 |    0 |
| hal/mbed_lp_ticker_api.o       |    48 |     0 |   24 |
| hal/mbed_sleep_manager.o       |   208 |     0 |    2 |
| hal/mbed_ticker_api.o          |   660 |     0 |    0 |
| hal/mbed_us_ticker_api.o       |    48 |     0 |   24 |
| platform/ATCmdParser.o         |    32 |     0 |    0 |
| platform/FileBase.o            |    96 |     0 |   56 |
| platform/FilePath.o            |   100 |     0 |    0 |
| platform/mbed_alloc_wrappers.o |    10 |     0 |    0 |
| platform/mbed_assert.o         |    84 |     0 |    0 |
| platform/mbed_board.o          |   600 |     0 |    0 |
| platform/mbed_critical.o       |    54 |     0 |    0 |
| platform/mbed_error.o          |    40 |     0 |    1 |
| platform/mbed_retarget.o       |   400 |     0 |  124 |
| platform/mbed_rtc_time.o       |   156 |     0 |    0 |
| platform/mbed_wait_api_rtos.o  |    56 |     0 |    0 |
| rtos/Mutex.o                   |   192 |     0 |    0 |
| rtos/RtosTimer.o               |    24 |     0 |    0 |
| rtos/TARGET_CORTEX             | 10866 |   180 | 6432 |
| rtos/Thread.o                  |    68 |     0 |    0 |
| targets/TARGET_NORDIC          |  7114 |   120 |  917 |
| Subtotals                      | 58835 |   976 | 9161 |
+--------------------------------+-------+-------+------+
Total Static RAM memory (data + bss): 10137 bytes
Total Flash memory (text + data): 59811 bytes

real    0.066505842s
```